Reorganize modules as @stjepang commented
