+++ src/bin/bench.rs
@@ -5,9 +5,8 @@ use std::sync::Mutex;
 use std::sync::mpsc::channel;
 use std::time::Duration;
 
-use crossbeam::utils::scoped::scope;
-use crossbeam::sync::MsQueue;
-use crossbeam::sync::SegQueue;
+use crossbeam::scope;
+use crossbeam::queue::{MsQueue, SegQueue};
 
 use extra_impls::mpsc_queue::Queue as MpscQueue;
